Issue#14を/sampleRouteMap下に移動してコンフリクト解消
