+++ frontend/package.json
@@ -20,7 +20,7 @@
     "next": "15.5.4",
     "react": "19.1.0",
     "react-dom": "19.1.0",
-    "react-leaflet": "^5.0.0"
+    "react-leaflet": "^5.0.0",
     "react-icons": "^5.5.0"
   },
   "devDependencies": {
